requirements.txt
# pigeon inference
ultralytics
python-vlc
ncnn
numpy==1.26.4

# notifications
python-telegram-bot

# development/production hosting
flask
waitress

# authentication
flask_limiter
passlib
flask_wtf
wtforms
bcrypt==4.0.1

# authentication/settings
redis

# misc
python-dotenv
requests